[2006-07-10] release for version 1.1.6

+++ scripts/01-sfk-selftest-win.bat
@@ -4,7 +4,7 @@ IF "%1"=="" GOTO xerr01
 IF ERRORLEVEL 1 GOTO xerr02
 
 cd ..
-rmdir /S /Q tmp-selftest
+IF EXIST tmp-selftest rmdir /S /Q tmp-selftest
 md tmp-selftest
 cd tmp-selftest
 xcopy ..\testfiles testfiles /S /H /I /R /K /Y >nul
